+++ src/python/pants/bin/BUILD
@@ -20,9 +20,9 @@ python_library(
     'src/python/pants/engine/legacy:address_mapper',
     'src/python/pants/engine/legacy:graph',
     'src/python/pants/engine/legacy:parser',
+    'src/python/pants/engine/legacy:round_engine',
     'src/python/pants/engine:build_files',
     'src/python/pants/engine:fs',
-    'src/python/pants/engine:legacy_engine',
     'src/python/pants/engine:mapper',
     'src/python/pants/engine:native',
     'src/python/pants/engine:parser',
