+++ src/python/pants/engine/legacy/BUILD
@@ -4,8 +4,24 @@
 python_tests(
   name='tests',
   dependencies=[
+    ':address_mapper',
     ':graph',
+    ':parser',
+    ':round_engine',
+    ':structs',
+    'src/python/pants/base:exceptions',
+    'src/python/pants/base:specs',
+    'src/python/pants/build_graph',
+    'src/python/pants/engine:fs',
+    'src/python/pants/engine:nodes',
+    'src/python/pants/engine:parser',
+    'src/python/pants/goal',
+    'src/python/pants/option',
+    'src/python/pants/task',
+    'src/python/pants/testutil/engine:engine_test_base',
     'src/python/pants/testutil:test_base',
+    'src/python/pants/util:contextutil',
+    'src/python/pants/util:dirutil',
   ],
   tags = {"type_checked"},
 )
@@ -15,13 +31,10 @@ python_library(
   sources=['address_mapper.py'],
   dependencies=[
     'src/python/pants/base:build_file',
+    'src/python/pants/base:exceptions',
     'src/python/pants/base:specs',
     'src/python/pants/build_graph',
-    'src/python/pants/engine:build_files',
-    'src/python/pants/engine:fs',
-    'src/python/pants/engine:mapper',
-    'src/python/pants/engine:nodes',
-    'src/python/pants/engine:selectors',
+    'src/python/pants/engine:addresses',
     'src/python/pants/util:dirutil',
   ],
   tags = {"partially_type_checked"},
@@ -33,25 +46,13 @@ python_library(
   dependencies=[
     ':structs',
     'src/python/pants/base:build_file_target_factory',
+    'src/python/pants/base:exceptions',
     'src/python/pants/base:parse_context',
-    'src/python/pants/engine:mapper',
+    'src/python/pants/build_graph',
     'src/python/pants/engine:objects',
     'src/python/pants/engine:parser',
-    'src/python/pants/util:memo',
-  ],
-  tags = {"partially_type_checked"},
-)
-
-python_library(
-  name='options_parsing',
-  sources=['options_parsing.py'],
-  dependencies=[
-    '3rdparty/python:dataclasses',
-    'src/python/pants/build_graph',
-    'src/python/pants/engine:fs',
-    'src/python/pants/engine:objects',
     'src/python/pants/option',
-    'src/python/pants/util:logging',
+    'src/python/pants/util:memo',
   ],
   tags = {"partially_type_checked"},
 )
@@ -63,13 +64,14 @@ python_library(
     '3rdparty/python:dataclasses',
     'src/python/pants/base:deprecated',
     'src/python/pants/build_graph',
+    'src/python/pants/engine:addressable',
     'src/python/pants/engine:fs',
-    'src/python/pants/engine:nodes',
+    'src/python/pants/engine:objects',
     'src/python/pants/engine:struct',
     'src/python/pants/engine:unions',
     'src/python/pants/source',
+    'src/python/pants/util:collections',
     'src/python/pants/util:contextutil',
-    'src/python/pants/util:memo',
     'src/python/pants/util:meta',
     'src/python/pants/util:objects',
   ],
@@ -81,21 +83,32 @@ python_library(
   sources=['graph.py'],
   dependencies=[
     '3rdparty/python:dataclasses',
-    ':address_mapper',
     ':structs',
     'src/python/pants/base:exceptions',
     'src/python/pants/base:parse_context',
+    'src/python/pants/base:specs',
     'src/python/pants/build_graph',
-    'src/python/pants/engine:build_files',
-    'src/python/pants/engine:mapper',
-    'src/python/pants/engine:objects',
+    'src/python/pants/engine:addresses',
+    'src/python/pants/engine:collection',
+    'src/python/pants/engine:fs',
     'src/python/pants/engine:parser',
+    'src/python/pants/engine:rules',
     'src/python/pants/engine:selectors',
-    'src/python/pants/engine:unions',
     'src/python/pants/option',
-    'src/python/pants/scm/subsystems:changed',
     'src/python/pants/source',
-    'src/python/pants/util:dirutil',
+    'src/python/pants/util:meta',
+    'src/python/pants/util:ordered_set',
+  ],
+  tags = {"partially_type_checked"},
+)
+
+python_library(
+  name = 'round_engine',
+  sources = ['round_engine.py'],
+  dependencies = [
+    'src/python/pants/base:exceptions',
+    'src/python/pants/base:workunit',
+    'src/python/pants/goal',
     'src/python/pants/util:ordered_set',
   ],
   tags = {"partially_type_checked"},
