+++ src/python/pants/init/BUILD
@@ -15,9 +15,10 @@ python_library(
     'src/python/pants/binaries',
     'src/python/pants/build_graph',
     'src/python/pants/core_tasks',
+    'src/python/pants/engine/internals:graph',
+    'src/python/pants/engine/internals:options_parsing',
     'src/python/pants/engine/legacy:address_mapper',
     'src/python/pants/engine/legacy:graph',
-    'src/python/pants/engine/legacy:options_parsing',
     'src/python/pants/engine/legacy:parser',
     'src/python/pants/engine/legacy:structs',
     'src/python/pants/engine:build_files',
