+++ tests/python/pants_test/engine/integration/test_goal_rule_integration.py
